fix(struct): typescript signature (#317)
The type definition for `struct` is out of date.

and does not include the `defaultProps` type definition. . Both of these are now fixed.

### Extend type signature
The extend method also does not have the correct type, as we can actually pass `StructOptions` to extend rather than a simple string. For example, this is currently giving the error:
``js
const a = t.struct({ foo: t.string });
const b = a.extend({ bar: t.string }, { name: 'b', strict: true }); // Expecting a simple string here, not StructProps
``

### defaultProps missing from StructOptions
`defaultProps` is not present in the StructOptions, this is incorrect. For example, this throws an error:
```js
const a = t.struct({ foo: t.string }, { defaultProps: { foo: 'bar' }}); // defaultProps is not a valid key
```