Fix: Enums.is() with an array value should always be false
Javascript oddity: `true === {p: 'v'}.hasOwnProperty(['p'])`.

In ancient times, the checked value was also ensured to be a String.
This was removed with #94 (`Enums` supports numbers).

Note: With this fix, valid `Enums` values are restricted to the
officially supported types `Number` and `String`. The previous
implementation could have worked also for `Symbol`.